only move configs around
I had a bug because sometimes I was moving layers around, instead of
copies of layers or their configs. This meant that sometimes a mutation
on one layer also mutated a second layer, simply because one was just a
pointer to another.

To fix this, I just do a lot of copying of lists and dicts to try and
avoid pointers wherever I can.