separate h and w dimensions
This is a simple generalisation away from square filters for conv and pool layers. It adds more parameters, but not a whole lot more.

Now for kernel size, pool size, and stride dimensions, we have a horizontal and a width parameter. So we can evolve all sorts of filters that were not previously possible.