introduce cache
This uses the relevant parts of each genotype to make and md5 hash. I then use this as a unique ID for that genotype, and store the vital stats for that genotype in a dict that serves as a cache. This way I don't need to re-train networks I've already seen.

The caches is rebuilt if lineages are loaded from .pkl files.