add max_time argument to evolve()
this allows me to kill off anything with a training time above max_time.

It's an alternative to using kill_slow(), because it's hard with kill_slow() to strike the balance between selecting too hard on time (at the cost of accuracy) versus allowing the training time to get out of hand.

The thinking here is that one can evolve populations with increasing levels of max_time, to sequentially allow for the evolution of more complex networks.